+++ kafka_bigquery_plugins/connector.properties
@@ -17,20 +17,41 @@ name=bigquery-connector
 connector.class=com.wepay.kafka.connect.bigquery.BigQuerySinkConnector
 tasks.max=10
 
-topics=metrics.json
+security.protocol=SASL_SSL
+sasl.mechanism=PLAIN
+request.timeout.ms=20000
+retry.backoff.ms=500
+ssl.endpoint.identification.algorithm=https
+bootstrap.servers=pkc-42n01.us-west-2.aws.devel.cpdev.cloud:9092
+
+sasl.jaas.config=org.apache.kafka.common.security.plain.PlainLoginModule required username\="7YZ3LUXXUDEEWTJX" password\="BW4aNjwoOtY8L6BcoPTv25NKDxioJ0+/4mml751org9fKbihtPPO7zCd/Y2hq4SB";
+
+
+consumer.ssl.endpoint.identification.algorithm=https
+consumer.sasl.mechanism=PLAIN
+consumer.request.timeout.ms=20000
+consumer.retry.backoff.ms=500
+consumer.sasl.jaas.config=org.apache.kafka.common.security.plain.PlainLoginModule required username\="7YZ3LUXXUDEEWTJX" password\="BW4aNjwoOtY8L6BcoPTv25NKDxioJ0+/4mml751org9fKbihtPPO7zCd/Y2hq4SB";
+consumer.security.protocol=SASL_SSL
+​
+producer.ssl.endpoint.identification.algorithm=https
+producer.sasl.mechanism=PLAIN
+producer.request.timeout.ms=20000
+producer.retry.backoff.ms=500
+producer.sasl.jaas.config=org.apache.kafka.common.security.plain.PlainLoginModule required username\="7YZ3LUXXUDEEWTJX" password\="BW4aNjwoOtY8L6BcoPTv25NKDxioJ0+/4mml751org9fKbihtPPO7zCd/Y2hq4SB";
+producer.security.protocol=SASL_SSL
+
+
+topics=metrics.json.kafka
 sanitizeTopics=true
 sanitizeFieldNames=true
 
 autoCreateTables=false
 autoUpdateSchemas=false
 
-
-# schemaRetriever=com.wepay.kafka.connect.bigquery.schemaregistry.schemaretriever.SchemaRegistrySchemaRetriever
-# schemaRegistryLocation=pkc-glad-giraffe.gcp.priv.cpdev.cloud\:8081
-
 bufferSize=100000
 maxWriteSize=10000
-tableWriteWait=1000
+tableWriteWait=100
 
 ########################################### Fill me in! ###########################################
 # The name of the BigQuery project to write to
