+++ kafka_bigquery_plugins/local_connector.properties
@@ -17,7 +17,7 @@ name=bigquery-connector
 connector.class=com.wepay.kafka.connect.bigquery.BigQuerySinkConnector
 tasks.max=10
 
-topics=test
+topics=metrics_json_kafka
 sanitizeTopics=true
 sanitizeFieldNames=true
