+++ kcbq-connector/src/main/java/com/wepay/kafka/connect/bigquery/preprocess/FieldNameSanitizer.java
@@ -1,0 +1,31 @@
+package com.wepay.kafka.connect.bigquery.preprocess;
+
+import java.util.Map;
+import java.util.stream.Collectors;
+
+public class FieldNameSanitizer {
+
+    private static String sanitizeName(String name) {
+        name = name.replaceAll("^[^a-zA-Z]+", "");
+        return name.replaceAll("[^a-zA-Z0-9_]", "_");
+    }
+
+
+    // Big Query specifies field name must begin with a letter or underscore and can only contain
+    // letters, numbers, and underscores
+    public static Map<String,Object> replaceInvalidKeys(Map<String, Object> map) {
+
+        return map.entrySet().stream().collect(Collectors.toMap(
+                (entry) -> {
+                    return sanitizeName((String) (entry.getKey()));
+                },
+                (entry) -> {
+                    if (entry.getValue() instanceof Map) {
+                        return replaceInvalidKeys((Map) entry.getValue());
+                    }
+                    return entry.getValue();
+
+                }
+        ));
+    }
+}
