--- kcbq-connector/src/main/java/com/wepay/kafka/connect/bigquery/transformer/FieldNameSanitizer.java
@@ -1,5 +1,0 @@
-package com.wepay.kafka.connect.bigquery.transformer;
-
-public class FieldNameSanitizer {
-
-}
