+++ kcbq-connector/src/test/java/com/wepay/kafka/connect/bigquery/preprocess/FieldNameSanitizerTest.java
@@ -1,0 +1,4 @@
+package com.wepay.kafka.connect.bigquery.preprocess;
+
+public class FieldNameSanitizerTest {
+}
