+++ kcbq-connector/src/test/java/com/wepay/kafka/connect/bigquery/preprocess/SinkRecordFilterTest.java
@@ -1,0 +1,12 @@
+package com.wepay.kafka.connect.bigquery.preprocess;
+
+
+import org.junit.Test;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.junit.Assert.assertEquals;
+
+public class SinkRecordFilterTest {
+}
